refacto(command-palette): separate into components
