+++ src/components/CommandPalette/actions.ts
@@ -1,0 +1,160 @@
+import {
+  Server,
+  Users,
+  Building,
+  Settings,
+  LogOut,
+  Plus,
+  Grid3X3,
+  List,
+  History,
+  Shield,
+  Plug
+} from 'lucide-vue-next';
+import type { CommandAction } from './types';
+
+export const createActions = (
+  navigateTo: (path: string) => void,
+  handleLogout: () => void,
+  switchView: (view: 'table' | 'card') => void,
+  isAdmin: boolean
+): CommandAction[] => [
+  {
+    group: 'command_palette.groups.navigation',
+    label: 'command_palette.dashboard',
+    description: 'command_palette.dashboard_desc',
+    icon: Grid3X3,
+    fn: () => navigateTo('/'),
+    shortcut: 'D',
+    path: '/',
+  },
+  {
+    group: 'command_palette.groups.navigation',
+    label: 'command_palette.servers',
+    description: 'command_palette.servers_desc',
+    icon: Server,
+    fn: () => navigateTo('/servers'),
+    shortcut: 'S',
+    path: '/servers',
+  },
+  {
+    group: 'command_palette.groups.navigation',
+    label: 'command_palette.users',
+    description: 'command_palette.users_desc',
+    icon: Users,
+    fn: () => navigateTo('/users'),
+    shortcut: 'U',
+    path: '/users',
+  },
+  {
+    group: 'command_palette.groups.navigation',
+    label: 'command_palette.rooms',
+    description: 'command_palette.rooms_desc',
+    icon: Building,
+    fn: () => navigateTo('/rooms'),
+    shortcut: 'R',
+    path: '/rooms',
+  },
+  {
+    group: 'command_palette.groups.navigation',
+    label: 'command_palette.ups',
+    description: 'command_palette.ups_desc',
+    icon: Plug,
+    fn: () => navigateTo('/ups'),
+    shortcut: 'P',
+    path: '/ups',
+  },
+  {
+    group: 'command_palette.groups.views',
+    label: 'command_palette.table_view',
+    description: 'command_palette.table_view_desc',
+    icon: List,
+    fn: () => switchView('table'),
+    onlyInUsers: true,
+  },
+  {
+    group: 'command_palette.groups.views',
+    label: 'command_palette.card_view',
+    description: 'command_palette.card_view_desc',
+    icon: Grid3X3,
+    fn: () => switchView('card'),
+    onlyInUsers: true,
+  },
+  {
+    group: 'command_palette.groups.create',
+    label: 'command_palette.create_server',
+    description: 'command_palette.create_server_desc',
+    icon: Plus,
+    fn: () => navigateTo('/servers/create'),
+    shortcut: '⌘+S',
+  },
+  {
+    group: 'command_palette.groups.create',
+    label: 'command_palette.create_room',
+    description: 'command_palette.create_room_desc',
+    icon: Plus,
+    fn: () => navigateTo('/rooms/create'),
+    shortcut: '⌘+R',
+  },
+  {
+    group: 'command_palette.groups.create',
+    label: 'command_palette.create_ups',
+    description: 'command_palette.create_ups_desc',
+    icon: Plus,
+    fn: () => navigateTo('/ups/create'),
+    shortcut: '⌘+P',
+  },
+  {
+    group: 'command_palette.groups.admin',
+    label: 'command_palette.admin_users',
+    description: 'command_palette.admin_users_desc',
+    icon: Shield,
+    fn: () => navigateTo('/admin/users'),
+    adminOnly: true,
+  },
+  {
+    group: 'command_palette.groups.admin',
+    label: 'command_palette.admin_roles',
+    description: 'command_palette.admin_roles_desc',
+    icon: Shield,
+    fn: () => navigateTo('/admin/roles'),
+    adminOnly: true,
+  },
+  {
+    group: 'command_palette.groups.admin',
+    label: 'command_palette.admin_history',
+    description: 'command_palette.admin_history_desc',
+    icon: History,
+    fn: () => navigateTo('/admin/history'),
+    adminOnly: true,
+  },
+  {
+    group: 'command_palette.groups.general',
+    label: 'command_palette.settings',
+    description: 'command_palette.settings_desc',
+    icon: Settings,
+    fn: () => navigateTo('/settings'),
+  },
+  {
+    group: 'command_palette.groups.general',
+    label: 'command_palette.profile',
+    description: 'command_palette.profile_desc',
+    icon: Users,
+    fn: () => navigateTo('/profile'),
+  },
+  {
+    group: 'command_palette.groups.general',
+    label: 'command_palette.changelog',
+    description: 'command_palette.changelog_desc',
+    icon: History,
+    fn: () => navigateTo('/changelog'),
+  },
+  {
+    group: 'command_palette.groups.general',
+    label: 'command_palette.logout',
+    description: 'command_palette.logout_desc',
+    icon: LogOut,
+    fn: () => handleLogout(),
+    shortcut: '⌘+⇧+Q',
+  },
+];
