+++ src/components/CommandPalette/types.ts
@@ -1,0 +1,11 @@
+export interface CommandAction {
+  group: string;
+  label: string;
+  description?: string;
+  icon: any;
+  fn: () => void;
+  shortcut?: string;
+  path?: string;
+  onlyInUsers?: boolean;
+  adminOnly?: boolean;
+}
